+++ lib/modules/state/state_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: state
-// Updated at: 2026-08-20 20:26:05
-// Commit: feat(state): integrate offline caching for state state
+// Updated at: 2026-08-20 22:16:12
+// Commit: perf(state): optimize memory consumption in state list viewport
 
 class stateComponentState {
-  final String commitRef = '3780';
-  final String lastMessage = 'feat(state): integrate offline caching for state state';
-  final DateTime timestamp = DateTime.parse('2026-08-20T20:26:05+05:00');
+  final String commitRef = '3786';
+  final String lastMessage = 'perf(state): optimize memory consumption in state list viewport';
+  final DateTime timestamp = DateTime.parse('2026-08-20T22:16:12+05:00');
 }